마이페이지에 좋아요누른거 저장된 레시피에 뜨게 되는거 확인, 눌러봤을 때 상세 내용 제대로 나오는거 확인, 좋아요같은거 기능 처리하는 과정에서 다 따로 나눠서 관리하라고 해서 새로 생긴게 좀 있음

+++ BACKEND/RECIPICK-PROJECT/src/main/java/SITE/RECIPICK/RECIPICK_PROJECT/controller/admin/AdminController.java
@@ -38,8 +38,7 @@ import org.springframework.web.bind.annotation.RestController;
  * 비즈니스 로직 수행 - 현재 로그인한 관리자 식별은 Spring Security Authentication + UserRepository로 처리
  *
  * <p>⚠️ 주의 - "임시 하드코딩된 사용자 ID"는 사용하지 않음. - 신고 등록 등 "신고자 ID"가 필요한 곳에서는 Authentication의 이름(email)로
- * DB
- * 조회 → userId 추출.
+ * DB 조회 → userId 추출.
  */
 @RestController
 @RequestMapping("/admin")
@@ -181,7 +180,7 @@ public class AdminController {
    */
   @PatchMapping("/reports/{id}")
   @Operation(summary = "신고 처리", description = "action=ACCEPT|REJECT")
-  public void moderate(@PathVariable Long id, @RequestBody ReportModerateRequest req) {
+  public void moderate(@PathVariable Integer id, @RequestBody ReportModerateRequest req) {
     svc.moderate(id, req);
   }
